Defined Query.__ne__ as the opposite of Query.__eq__. Fixed RangeMixin.__eq__.
Bizarre Python thing I learned today: __ne__() does not automatically return the opposite of __eq__, you actually have to define it. What I don't understand is why the fact that __ne__ wasn't defined masked this problem on Windows and Mac OS X but not on Ubuntu. There might be a Python platform bug here.

Anyway, I fixed the problem that was masking the error on Mac and Windows (lack of __ne__), and then fixed the error (RangeMixin.__eq__ wasn't comparing constantscore).

Fixes issue #108.