DATETIME.to_text - extend type support to text_type (unicode/str)
If whoosh parses a query for a DATETIME field, it makes quite some effort to
make sense out of numeric strings, so something like 2011 or
20111231235959999999 is parsed as one would expect it.

In this changeset, I modified method .to_text (which is used at indexing time)
to also try to make sense out of numeric strings (in about the same way).

It calls floor() so that 2011 means the same as 201101010000000000 (floor
makes most sense, 42 and 42.0000000 is usually considered same thing, so,
following same logic, if you just say 2011, but we need some specific
datetime, that shall be the very first moment in that year).