Fixed a bug in MultiReader._merge_terms() with empty iterators. Fixes issue #267.
This took a long time to realize, but I was calling next() without catching
StopIterator when getting the first terms from each iterator. If the iterator was
empty, it would raise StopIterator. Since it wasn't caught, the exception propagated
upwards, and then since the method itself is a generator, the StopIterator would be
silently swallowed and the generator would exit, making it seem like the method would
just return for no reason.

--HG--
branch : 2.4x